requirements.txt
streamlit
tensorflow
keras
opencv-python
pillow
numpy
streamlit-webrtc
gdown
